fix frontend, rules, and auth

+++ app/Controllers/Backend/Bantuans.php
@@ -2,6 +2,8 @@
 
 namespace App\Controllers\Backend;
 
+use App\Libraries\GetDesa;
+use App\Libraries\GetUser;
 use CodeIgniter\API\ResponseTrait;
 use CodeIgniter\RESTful\ResourceController;
 
@@ -10,11 +12,13 @@ class Bantuans extends ResourceController
     use ResponseTrait;
     protected $modelName = 'App\Models\BantuansModel';
     protected $format    = 'json';
+    protected $db, $user, $desa;
 
-    protected $db;
     public function __construct()
     {
         $this->db = \Config\Database::connect();
+        $this->user = new GetUser();
+        $this->desa = new GetDesa();
     }
 
     /**
@@ -24,7 +28,10 @@ class Bantuans extends ResourceController
      */
     public function index()
     {
-        $data = $this->model->findAll();
+        $currUser = $this->user->currLogin();
+        if ($currUser['desa'] != 'admin') return $this->fail('desa not allowed');
+
+        $data = $this->model->orderBy('id', 'DESC')->findAll();
         return $this->respond($data);
     }
 
@@ -37,6 +44,13 @@ class Bantuans extends ResourceController
     {
         $data = $this->model->find($id);
         if (!$data) return $this->failNotFound('no data found');
+
+        $currUser = $this->user->currLogin();
+        $currDesa = $this->desa->currDesa($data['id_desas']);
+        if ($currUser['desa'] != 'admin') {
+            if ($currUser['desa'] != $currDesa['desa']) return $this->fail('desa not allowed');
+        }
+
         return $this->respond($data);
     }
 
@@ -47,6 +61,12 @@ class Bantuans extends ResourceController
      */
     public function create()
     {
+        $currUser = $this->user->currLogin();
+        $currDesa = $this->desa->currDesa($this->request->getVar('id_desas'));
+        if ($currUser['desa'] != 'admin') {
+            if ($currUser['desa'] != $currDesa['desa']) return $this->fail('desa not allowed');
+        }
+
         helper(['form']);
 
         $rules = $this->model->myValidationRules;
@@ -93,6 +113,12 @@ class Bantuans extends ResourceController
     {
         $findData = $this->model->find($id);
         if (!$findData) return $this->failNotFound('no data found');
+
+        $currUser = $this->user->currLogin();
+        $currDesa = $this->desa->currDesa($findData('id_desas'));
+        if ($currUser['desa'] != 'admin') {
+            if ($currUser['desa'] != $currDesa['desa']) return $this->fail('desa not allowed');
+        }
 
         helper(['form']);
 
@@ -148,6 +174,12 @@ class Bantuans extends ResourceController
         $findData = $this->model->find($id);
         if (!$findData) return $this->failNotFound('no data found');
 
+        $currUser = $this->user->currLogin();
+        $currDesa = $this->desa->currDesa($findData('id_desas'));
+        if ($currUser['desa'] != 'admin') {
+            if ($currUser['desa'] != $currDesa['desa']) return $this->fail('desa not allowed');
+        }
+
         if (file_exists('img/bantuan/' . $findData['foto'])) {
             unlink('img/bantuan/' . $findData['foto']);
         }
@@ -200,10 +232,68 @@ class Bantuans extends ResourceController
             $where = null;
         }
 
+        $currUser = $this->user->currLogin();
+        if ($currUser['desa'] != 'admin') {
+            $currDesa = $currUser['desa'];
+            $desasCrr = $this->db->table('desas')
+                ->getWhere("desa = '$currDesa'")
+                ->getResultArray();
+            $desaId = $desasCrr[0]['id'];
+            $where = "id_desas = '$desaId'";
+        }
+
         $bantuans = $this->db->table('bantuans')
             ->orderBy('bantuans.id', 'DESC')
             ->getWhere($where, $limit, $offset)
             ->getResultArray();
+
+        $provinsis = $this->db->table('provinsis')
+            ->get()
+            ->getResultArray();
+
+        foreach ($bantuans as $key => $bantuan) {
+            foreach ($provinsis as $provinsi) {
+                if ($bantuan['id_provinsis'] == $provinsi['id']) {
+                    $bantuans[$key]['provinsi'] = $provinsi['provinsi'];
+                }
+            }
+        }
+
+        $kabupatens = $this->db->table('kabupatens')
+            ->get()
+            ->getResultArray();
+
+        foreach ($bantuans as $key => $bantuan) {
+            foreach ($kabupatens as $kabupaten) {
+                if ($bantuan['id_kabupatens'] == $kabupaten['id']) {
+                    $bantuans[$key]['kabupaten'] = $kabupaten['kabupaten'];
+                }
+            }
+        }
+
+        $kecamatans = $this->db->table('kecamatans')
+            ->get()
+            ->getResultArray();
+
+        foreach ($bantuans as $key => $bantuan) {
+            foreach ($kecamatans as $kecamatan) {
+                if ($bantuan['id_kecamatans'] == $kecamatan['id']) {
+                    $bantuans[$key]['kecamatan'] = $kecamatan['kecamatan'];
+                }
+            }
+        }
+
+        $desas = $this->db->table('desas')
+            ->get()
+            ->getResultArray();
+
+        foreach ($bantuans as $key => $bantuan) {
+            foreach ($desas as $desa) {
+                if ($bantuan['id_desas'] == $desa['id']) {
+                    $bantuans[$key]['desa'] = $desa['desa'];
+                }
+            }
+        }
 
         $provinsis = $this->db->table('provinsis')
             ->get()
