+++ app/Controllers/Backend/DataWilayahs.php
@@ -2,6 +2,8 @@
 
 namespace App\Controllers\Backend;
 
+use App\Libraries\GetDesa;
+use App\Libraries\GetUser;
 use CodeIgniter\API\ResponseTrait;
 use CodeIgniter\RESTful\ResourceController;
 
@@ -10,11 +12,13 @@ class DataWilayahs extends ResourceController
     use ResponseTrait;
     protected $modelName = 'App\Models\DataWilayahsModel';
     protected $format    = 'json';
+    protected $db, $user, $desa;
 
-    protected $db;
     public function __construct()
     {
         $this->db = \Config\Database::connect();
+        $this->user = new GetUser();
+        $this->desa = new GetDesa();
     }
 
     /**
@@ -24,6 +28,9 @@ class DataWilayahs extends ResourceController
      */
     public function index()
     {
+        $currUser = $this->user->currLogin();
+        if ($currUser['desa'] != 'admin') return $this->fail('desa not allowed');
+
         $data = $this->model->orderBy('id', 'DESC')->findAll();
         return $this->respond($data);
     }
@@ -37,6 +44,13 @@ class DataWilayahs extends ResourceController
     {
         $data = $this->model->find($id);
         if (!$data) return $this->failNotFound('no data found');
+
+        $currUser = $this->user->currLogin();
+        $currDesa = $this->desa->currDesa($data['id_desas']);
+        if ($currUser['desa'] != 'admin') {
+            if ($currUser['desa'] != $currDesa['desa']) return $this->fail('desa not allowed');
+        }
+
         return $this->respond($data);
     }
 
@@ -47,6 +61,12 @@ class DataWilayahs extends ResourceController
      */
     public function create()
     {
+        $currUser = $this->user->currLogin();
+        $currDesa = $this->desa->currDesa($this->request->getVar('id_desas'));
+        if ($currUser['desa'] != 'admin') {
+            if ($currUser['desa'] != $currDesa['desa']) return $this->fail('desa not allowed');
+        }
+
         helper(['form']);
 
         $rules = $this->model->myValidationRules;
@@ -193,6 +213,12 @@ class DataWilayahs extends ResourceController
         $findData = $this->model->find($id);
         if (!$findData) return $this->failNotFound('no data found');
 
+        $currUser = $this->user->currLogin();
+        $currDesa = $this->desa->currDesa($findData('id_desas'));
+        if ($currUser['desa'] != 'admin') {
+            if ($currUser['desa'] != $currDesa['desa']) return $this->fail('desa not allowed');
+        }
+
         helper(['form']);
 
         $rules = $this->model->myValidationRules;
@@ -336,6 +362,12 @@ class DataWilayahs extends ResourceController
         $findData = $this->model->find($id);
         if (!$findData) return $this->failNotFound('no data found');
 
+        $currUser = $this->user->currLogin();
+        $currDesa = $this->desa->currDesa($findData('id_desas'));
+        if ($currUser['desa'] != 'admin') {
+            if ($currUser['desa'] != $currDesa['desa']) return $this->fail('desa not allowed');
+        }
+
         $this->model->delete($id);
         $response = [
             'status' => 200,
@@ -382,6 +414,16 @@ class DataWilayahs extends ResourceController
             }
         } else {
             $where = null;
+        }
+
+        $currUser = $this->user->currLogin();
+        if ($currUser['desa'] != 'admin') {
+            $currDesa = $currUser['desa'];
+            $desasCrr = $this->db->table('desas')
+                ->getWhere("desa = '$currDesa'")
+                ->getResultArray();
+            $desaId = $desasCrr[0]['id'];
+            $where = "id_desas = '$desaId'";
         }
 
         $data_wilayahs = $this->db->table('data_wilayahs')
